+++ packages/conf/__typetests__/index.tst.ts
@@ -22,7 +22,32 @@ expect({
 // pseudoLocale as an object with pseudolocale options
 expect({
   locales: ["en", "pseudo"],
-  pseudoLocale: { locale: "pseudo", prepend: "⟦ ", append: " ⟧", extend: 0.4 },
+  pseudoLocale: {
+    locale: "pseudo",
+    prepend: "⟦ ",
+    append: " ⟧",
+    extend: 0.4,
+    extendCharacter: ".",
+    override: "_",
+    rightToLeft: true,
+  },
+}).type.toBeAssignableTo<LinguiConfig>()
+
+// pseudoLocale as an array of objects with pseudolocale options
+expect({
+  locales: ["en", "pseudo-en", "pseudo-ar"],
+  pseudoLocale: [
+    {
+      locale: "pseudo-en",
+      prepend: "⟦ ",
+      append: " ⟧",
+      extend: 0.4,
+    },
+    {
+      locale: "pseudo-ar",
+      rightToLeft: true,
+    },
+  ],
 }).type.toBeAssignableTo<LinguiConfig>()
 
 // all props